interview_bot/interview-chatbot/backend/requirements.txt
openai
transformers
torch
spacy
flask
flask-cors
psycopg2-binary
pdfminer.six
python-docx
lightning
nltk
pandas
scikit-learn
